Makes changes per Punit review
