--- src/models/requests.js
@@ -1,5 +1,0 @@
-const db = require('./db/queries/requests');
-
-module.exports = {
-  add: db.add
-};
