+++ src/models/timeoffRequests.js
@@ -1,0 +1,5 @@
+const db = require('./db/queries/timeoffRequests');
+
+module.exports = {
+  add: db.add
+};
